Add safeguards to inconsistent New Relic metrics
Some of the items in application_summary from the New Relic API are
optional and may or may not be present. If they are not present, the
script silently fails and does not show any information because it
encounters an error.

This makes the inconsistent keys optional. If you do not have these keys
present in your application_summary output from New Relic they will show
up as "Undefined" in the output which is preferable over the plugin not
working.